requirements.txt
Django<3.3
poweradmin @ git+https://bitbucket.org/josir/poweradmin.git@django32
django-admin-tools==0.9.2
django-widget-tweaks==1.4.12
lxml==4.9.2
beautifulsoup4==4.11.2
mysqlclient==2.1.1
requests==2.28.2
python-stdnum==1.18
simhash==2.1.2
python-slugify==8.0.1
python-Levenshtein==0.20.9
crossrefapi==1.5.0
django-recaptcha==3.0.0
python-dateutil==2.8.2
jmespath==1.0.1
botocore==1.29.99
s3transfer==0.6.0
boto3==1.26.99
django-storages==1.13.2
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-appconf==1.0.5
django-select2==8.1.1
django-smart-selects==1.6.0